アプリケーションのメイン部分をPeerCastStation.App.dllに分割した
アプリケーションのメイン部分を実行ファイル起動後に別AppDomainで実行したいので、
PeerCastStation.exeに入っていたアプリケーションメイン部分をPeerCastStation.App.dllに分割した。
