+++ PeerCastStation/PeerCastStation.App/StandaloneApp.cs
@@ -1,0 +1,88 @@
+﻿using System;
+using System.Linq;
+using PeerCastStation.Core;
+using System.Threading;
+
+namespace PeerCastStation.App
+{
+  public class StandaloneApp
+    : Application
+  {
+    private PecaSettings settings = new PecaSettings(PecaSettings.DefaultFileName);
+    public override PecaSettings Settings {
+      get { return settings; }
+    }
+
+    public StandaloneApp()
+    {
+    }
+
+    protected override void DoSetup()
+    {
+      Console.CancelKeyPress += (sender, args) => {
+        args.Cancel = true;
+        Stop();
+      };
+      RegisteredWaitHandle registered = null;
+      registered = ThreadPool.RegisterWaitForSingleObject(killWaitHandle, (state, timedout) => {
+        Stop();
+        if (registered!=null) {
+          registered.Unregister(null);
+        }
+      }, null, Timeout.Infinite, true);
+      base.DoSetup();
+    }
+
+    protected override void DoCleanup()
+    {
+      base.DoCleanup();
+    }
+
+    static EventWaitHandle killWaitHandle;
+    static private bool CheckIsFirstInstance(ref EventWaitHandle wait_handle)
+    {
+      bool is_first_instance;
+      var event_name = System.Reflection.Assembly.GetEntryAssembly().Location
+        .Replace('\\', '/')+".kill";
+      try {
+        wait_handle = EventWaitHandle.OpenExisting(event_name);
+        is_first_instance = false;
+      }
+      catch (WaitHandleCannotBeOpenedException) {
+        wait_handle = new EventWaitHandle(false, EventResetMode.ManualReset, event_name);
+        is_first_instance = true;
+      }
+      return is_first_instance;
+    }
+
+    [STAThread]
+    public static void Run(string[] args)
+    {
+      AppDomain.CurrentDomain.UnhandledException += OnUnhandledException;
+      var first_instance = CheckIsFirstInstance(ref killWaitHandle);
+      if (args.Contains("-kill")) {
+        killWaitHandle.Set();
+        return;
+      }
+      if (!first_instance && !args.Contains("-multi")) {
+        return;
+      }
+      (new StandaloneApp()).Run();
+    }
+
+    private static void OnUnhandledException(object sender, UnhandledExceptionEventArgs args)
+    {
+      var dir = System.IO.Path.GetDirectoryName(PecaSettings.DefaultFileName);
+      System.IO.Directory.CreateDirectory(dir);
+      using (var file=System.IO.File.AppendText(System.IO.Path.Combine(dir, "exception.log"))) {
+        file.WriteLine("{0}: {1} (OS:{2}, CLR:{3})",
+          DateTime.Now,
+          AppSettingsReader.GetString("AgentName", "PeerCastStation"),
+          Environment.OSVersion,
+          Environment.Version);
+        file.WriteLine(args.ExceptionObject);
+      }
+    }
+
+  }
+}
