+++ PeerCastStation/PeerCastStation.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26403.7
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PeerCastStation.Core", "PeerCastStation.Core\PeerCastStation.Core.csproj", "{387996F5-E31D-4D92-BD86-8983C599F748}"
 EndProject
@@ -82,6 +82,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PeerCastStation.TS", "PeerCastStation.TS\PeerCastStation.TS.csproj", "{0F0123AE-E500-45CA-B69D-4FBDBB88B752}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PeerCastStation.CustomFilter", "PeerCastStation.CustomFilter\PeerCastStation.CustomFilter.csproj", "{E45E0A0A-4ACC-4B46-A4B4-E2D836BDE3DD}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PeerCastStation.App", "PeerCastStation.App\PeerCastStation.App.csproj", "{C8B4E4DC-BE2B-4FC5-94D2-A22FF754C9A7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -172,6 +174,10 @@ Global
 		{E45E0A0A-4ACC-4B46-A4B4-E2D836BDE3DD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E45E0A0A-4ACC-4B46-A4B4-E2D836BDE3DD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E45E0A0A-4ACC-4B46-A4B4-E2D836BDE3DD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C8B4E4DC-BE2B-4FC5-94D2-A22FF754C9A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C8B4E4DC-BE2B-4FC5-94D2-A22FF754C9A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C8B4E4DC-BE2B-4FC5-94D2-A22FF754C9A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C8B4E4DC-BE2B-4FC5-94D2-A22FF754C9A7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
